feat: Update task position handling to support fractional values
- Changed database schema to convert position fields from INTEGER to NUMERIC(10,2) for tasks, columns, and boards, allowing for fractional positioning.
- Implemented migration logic to alter existing position columns in PostgreSQL and handle schema checks for SQLite.
- Enhanced task copy functionality to insert copied tasks at fractional positions, ensuring correct order without disrupting existing tasks.
- Updated frontend logic to renumber tasks after copying, maintaining clean integer positions for user interface consistency.
- Improved task reordering logic to accommodate new position handling, ensuring reliable updates across columns.

+++ server/config/database.js
@@ -295,7 +295,7 @@ const CREATE_TABLES_SQL = `
       id TEXT PRIMARY KEY,
       title TEXT NOT NULL,
       project TEXT,
-      position INTEGER DEFAULT 0,
+      position NUMERIC(10,2) DEFAULT 0,
       created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
       updated_at DATETIME DEFAULT CURRENT_TIMESTAMP
     );
@@ -304,7 +304,7 @@ const CREATE_TABLES_SQL = `
       id TEXT PRIMARY KEY,
       boardId TEXT NOT NULL,
       title TEXT NOT NULL,
-      position INTEGER DEFAULT 0,
+      position NUMERIC(10,2) DEFAULT 0,
       is_finished BOOLEAN DEFAULT 0,
       is_archived BOOLEAN DEFAULT 0,
       created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
@@ -314,7 +314,7 @@ const CREATE_TABLES_SQL = `
 
     CREATE TABLE IF NOT EXISTS tasks (
       id TEXT PRIMARY KEY NOT NULL,
-      position INTEGER DEFAULT 0,
+      position NUMERIC(10,2) DEFAULT 0,
       title TEXT NOT NULL,
       description TEXT,
       ticket TEXT,
